+++ packages/lit/index.html
@@ -7,12 +7,12 @@
     <title>Vite + Lit + TS</title>
     <link rel="stylesheet" href="./src/index.css" />
     <script type="module" src="/src/my-element.ts"></script>
-    <script type="module" src="/src/components/my-text-input.ts"></script>
+    <script type="module" src="/src/components/lit/my-search-input.ts"></script>
   </head>
   <body>
     <my-element>
       <h1>Vite + Lit</h1>
-      <my-text-input></my-text-input>
+      <my-search-input></my-search-input>
     </my-element>
   </body>
 </html>
