+++ packages/lit/types/components/lit/my-search-input.d.ts
@@ -1,13 +1,11 @@
 import { LitElement } from 'lit';
-export declare class MyTextInput extends LitElement {
-    private _text;
+export declare class MySearchInput extends LitElement {
+    placeholder: string;
     render(): import("lit-html").TemplateResult<1>;
-    input: HTMLInputElement;
-    submitText(): void;
     static styles: import("lit").CSSResult;
 }
 declare global {
     interface HTMLElementTagNameMap {
-        'my-clock': MyTextInput;
+        'my-search-input': MySearchInput;
     }
 }
